feat: 2302 극장 좌석(DP)
