+++ wonu606 /백준/Silver/2302. 극장 좌석/README.md
