--- 백준/Gold/2230. 수 고르기/README.md
@@ -1,26 +1,0 @@
-# [Gold V] 수 고르기 - 2230 
-
-[문제 링크](https://www.acmicpc.net/problem/2230) 
-
-### 성능 요약
-
-메모리: 28100 KB, 시간: 360 ms
-
-### 분류
-
-정렬, 두 포인터
-
-### 문제 설명
-
-<p>N개의 정수로 이루어진 수열 A[1], A[2], …, A[N]이 있다. 이 수열에서 두 수를 골랐을 때(같은 수일 수도 있다), 그 차이가 M 이상이면서 제일 작은 경우를 구하는 프로그램을 작성하시오.</p>
-
-<p>예를 들어 수열이 {1, 2, 3, 4, 5}라고 하자. 만약 M = 3일 경우, 1 4, 1 5, 2 5를 골랐을 때 그 차이가 M 이상이 된다. 이 중에서 차이가 가장 작은 경우는 1 4나 2 5를 골랐을 때의 3이 된다.</p>
-
-### 입력 
-
- <p>첫째 줄에 두 정수 N, M이 주어진다. 다음 N개의 줄에는 차례로 A[1], A[2], …, A[N]이 주어진다.</p>
-
-### 출력 
-
- <p>첫째 줄에 M 이상이면서 가장 작은 차이를 출력한다. 항상 차이가 M이상인 두 수를 고를 수 있다.</p>
-
